Address the concerns of #345
This adds the ability for a user to set their own html resources,
intended for use by developers when debugging compatibility issues.

The API for this is as follows:

* The mine admin must enable this feature by setting
"head.allow.user.overrides = true" in any property file that
contributes to `WEB_PROPERTIES`.
* The user should then identify the key of the property they wish to
override. This is easily done by inspecting the head where each resource
has a comment indicating its key.
* The user should indicate their preferred value by running the
following js in the console:

```:js
$SERVICE.whoami().then(function(user) { user.setPreference(key, value)});
```

eg. to replace bootstrap with a different version, the user would run:

```:js
$SERVICE.whoami().then(function(user) {
    user.setPreference(
        'head.js.all.Ac',
        'http://cdn.intermine.org/js/backbone.js/0.9.9/backbone-min.js'
    );
});
```