Remove planetj CompressionFilter automatic compression of tomcat responses because this does not work with Tomcat8 (causes reponse delay until timeout).
If the mine operator wants compressed responses, they will need to configure Tomcat appropriately.
This is probably better since compression is not desired in all cases.
In InterMine's case, I should think that it's not enormously valuable and the performance bottleneck is usually cpu and database response time on the server.